Client keys loaded in run command
