+++ clients/desktop/src/client/mod.rs
@@ -63,20 +63,7 @@ pub struct NymClient {
 pub(crate) struct InputMessage(pub Destination, pub Vec<u8>);
 
 impl NymClient {
-    fn load_identity_keys(config_file: &Config) -> MixIdentityKeyPair {
-        let identity_keypair = PemStore::new(ClientPathfinder::new_from_config(&config_file))
-            .read_identity()
-            .expect("Failed to read stored identity key files");
-        println!(
-            "Public identity key: {}\n",
-            identity_keypair.public_key.to_base58_string()
-        );
-        identity_keypair
-    }
-
-    pub fn new(config: Config) -> Self {
-        let identity_keypair = Self::load_identity_keys(&config);
-
+    pub fn new(config: Config, identity_keypair: MixIdentityKeyPair) -> Self {
         NymClient {
             runtime: Runtime::new().unwrap(),
             config,
