+++ clients/desktop/src/commands/run.rs
@@ -14,9 +14,11 @@
 
 use crate::client::NymClient;
 use crate::commands::override_config;
-use crate::config::Config;
+use crate::config::{persistence::pathfinder::ClientPathfinder, Config};
 use clap::{App, Arg, ArgMatches};
 use config::NymConfig;
+use crypto::identity::MixIdentityKeyPair;
+use pemstore::pemstore::PemStore;
 
 pub fn command_args<'a, 'b>() -> clap::App<'a, 'b> {
     App::new("run")
@@ -55,6 +57,17 @@ pub fn command_args<'a, 'b>() -> clap::App<'a, 'b> {
         )
 }
 
+fn load_identity_keys(config_file: &Config) -> MixIdentityKeyPair {
+    let identity_keypair = PemStore::new(ClientPathfinder::new_from_config(&config_file))
+        .read_identity()
+        .expect("Failed to read stored identity key files");
+    println!(
+        "Public identity key: {}\n",
+        identity_keypair.public_key.to_base58_string()
+    );
+    identity_keypair
+}
+
 pub fn execute(matches: &ArgMatches) {
     let id = matches.value_of("id").unwrap();
 
@@ -63,5 +76,6 @@ pub fn execute(matches: &ArgMatches) {
             .expect("Failed to load config file");
 
     config = override_config(config, matches);
-    NymClient::new(config).run_forever();
+    let identity_keypair = load_identity_keys(&config);
+    NymClient::new(config, identity_keypair).run_forever();
 }
